test: statistics
