+++ annofabcli/statistics/subcommand_statistics.py
@@ -2,12 +2,12 @@ import argparse
 
 import annofabcli
 import annofabcli.common.cli
+import annofabcli.stat_visualization.merge_visualization_dir
 import annofabcli.statistics.list_annotation_count
 import annofabcli.statistics.list_by_date_user
 import annofabcli.statistics.list_cumulative_labor_time
 import annofabcli.statistics.list_labor_time_per_user
 import annofabcli.statistics.list_task_progress
-import annofabcli.stat_visualization.merge_visualization_dir
 import annofabcli.statistics.summarize_task_count
 import annofabcli.statistics.summarize_task_count_by_task_id
 import annofabcli.statistics.summarize_task_count_by_user
@@ -20,9 +20,9 @@ def parse_args(parser: argparse.ArgumentParser):
 
     # サブコマンドの定義
     annofabcli.statistics.list_annotation_count.add_parser(subparsers)
+    annofabcli.statistics.list_by_date_user.add_parser(subparsers)
     annofabcli.statistics.list_cumulative_labor_time.add_parser(subparsers)
     annofabcli.statistics.list_labor_time_per_user.add_parser(subparsers)
-    annofabcli.statistics.list_by_date_user.add_parser(subparsers)
     annofabcli.statistics.list_task_progress.add_parser(subparsers)
     annofabcli.statistics.summarize_task_count.add_parser(subparsers)
     annofabcli.statistics.summarize_task_count_by_task_id.add_parser(subparsers)
