+++ docs/command_reference/statistics/index.rst
@@ -17,20 +17,11 @@ Available Commands
 
 
    list_annotation_count
-                       
-   list_cumulative_labor_time
-                       
-   list_labor_time_per_user
-                       
    list_by_date_user   
+   list_cumulative_labor_time
+   list_labor_time_per_user
    list_task_progress  
-                       
    summarize_task_count
-                       
    summarize_task_count_by_task_id
-                       
    summarize_task_count_by_user
-                       
    visualize
-
-   merge_visualization_dir
