+++ tests/test_main.py
@@ -255,169 +255,6 @@ class TestProjectMember:
         )
 
 
-class TestStatistics:
-    def test_visualize(self):
-        output_dir = str(out_path / "statistics")
-        main(
-            [
-                "statistics",
-                "visualize",
-                "--project_id",
-                project_id,
-                "--task_query",
-                '{"status": "complete"}',
-                "--output_dir",
-                output_dir,
-                "--minimal",
-            ]
-        )
-
-    def test_merge_visualization_dir(self):
-        output_dir = str(out_path / "statistics/merge")
-        main(
-            [
-                "statistics",
-                "merge_visualization_dir",
-                "--dir",
-                str(data_path / "statistics/visualization-dir"),
-                "--output_dir",
-                output_dir,
-                "--minimal",
-            ]
-        )
-
-    def test_list_task_progress(self):
-        out_file = str(out_path / "task-progress.csv")
-        main(
-            [
-                "statistics",
-                "list_task_progress",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_cumulative_labor_time(self):
-        out_file = str(out_path / "cumulative-labor-time.csv")
-        main(
-            [
-                "statistics",
-                "list_cumulative_labor_time",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_annotation_count(self):
-        output_dir = str(out_path / "statistics-list-annotation-count")
-        main(
-            [
-                "statistics",
-                "list_annotation_count",
-                "--project_id",
-                project_id,
-                "--output_dir",
-                output_dir,
-                "--group_by",
-                "task_id",
-            ]
-        )
-        main(
-            [
-                "statistics",
-                "list_annotation_count",
-                "--project_id",
-                project_id,
-                "--output_dir",
-                output_dir,
-                "--group_by",
-                "input_data_id",
-            ]
-        )
-
-    def test_summarize_task_count(self):
-        out_file = str(out_path / "summariz-task-count.csv")
-        main(
-            [
-                "statistics",
-                "summarize_task_count",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-        out_file = str(out_path / "summariz-task-count2.csv")
-        main(
-            [
-                "statistics",
-                "summarize_task_count",
-                "--project_id",
-                project_id,
-                "--task_json",
-                str(data_path / "task.json"),
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_summarize_task_count_by_task_id(self):
-        out_file = str(out_path / "summarize_task_count_by_task_id.csv")
-        main(
-            [
-                "statistics",
-                "summarize_task_count_by_task_id",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_summarize_task_count_by_user(self):
-        out_file = str(out_path / "summarize_task_count_by_user.csv")
-        main(
-            [
-                "statistics",
-                "summarize_task_count_by_user",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_by_date_user(self):
-        out_file = str(out_path / "list_by_date_user.csv")
-        main(
-            [
-                "statistics",
-                "list_by_date_user",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_labor_time_per_user(self):
-        out_file = str(out_path / "list_labor_time_per_user.csv")
-        main(
-            [
-                "statistics",
-                "list_labor_time_per_user",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-
 class TestTaskHistory:
     command_name = "task_history"
